@@ -9,8 +9,8 @@
   },
   "private": true,
   "engines": {
-    "node": "12.16.2",
-    "yarn": "^2.0.0-rc.32"
+    "node": "12.17.0",
+    "yarn": "^2.0.0-rc.33"
   },
   "scripts": {
     "js": "rollup -c",
@@ -43,61 +43,60 @@
   },
   "devDependencies": {
     "#": "link:./src/filters",
-    "@11ty/eleventy": "^0.10.0",
+    "@11ty/eleventy": "^0.11.0",
     "@11ty/eleventy-plugin-rss": "^1.0.7",
     "@11ty/eleventy-plugin-syntaxhighlight": "^3.0.1",
-    "@babel/core": "^7.9.0",
-    "@babel/preset-env": "^7.9.5",
+    "@babel/core": "^7.10.2",
+    "@babel/preset-env": "^7.10.2",
     "@gerhobbelt/markdown-it-responsive": "^0.1.1-3",
-    "@rollup/plugin-commonjs": "^11.1.0",
-    "@rollup/plugin-node-resolve": "^7.1.3",
-    "@yarnpkg/pnpify": "^2.0.0-rc.21",
+    "@rollup/plugin-babel": "^5.0.2",
+    "@rollup/plugin-commonjs": "^12.0.0",
+    "@rollup/plugin-node-resolve": "^8.0.0",
+    "@yarnpkg/pnpify": "^2.0.0-rc.22",
     "accoutrement": "^2.1.3",
-    "autoprefixer": "^9.7.6",
-    "babel-jest": "^25.4.0",
+    "autoprefixer": "^9.8.0",
+    "babel-jest": "^26.0.1",
     "cascading-color-systems": "^0.1.0-beta.11",
     "chalk": "^4.0.0",
     "chokidar-cli": "^2.1.0",
     "core-js": "^3.6.5",
     "cssremedy": "^0.1.0-beta.2",
     "doxray": "^0.10.0",
-    "eslint": "^6.8.0",
+    "eslint": "^7.1.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-import-resolver-node": "^0.3.3",
     "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-jest": "^23.8.2",
+    "eslint-plugin-jest": "^23.13.2",
     "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-simple-import-sort": "^5.0.2",
+    "eslint-plugin-simple-import-sort": "^5.0.3",
     "fs-extra": "^9.0.0",
     "gulp": "^4.0.2",
     "gulp-responsive": "^3.0.1",
     "husky": "^4.2.5",
-    "jest": "^25.4.0",
-    "js-yaml": "^3.13.1",
-    "lint-staged": "^10.1.7",
+    "jest": "^26.0.1",
+    "js-yaml": "^3.14.0",
+    "lint-staged": "^10.2.7",
     "lodash": "^4.17.15",
-    "markdown-it": "^10.0.0",
+    "markdown-it": "^11.0.0",
     "markdown-it-footnote": "^3.0.2",
     "markdown-it-mark": "^3.0.0",
-    "mockdate": "^2.0.5",
-    "node-sass": "^4.13.1",
+    "mockdate": "^3.0.2",
     "npm-run-all": "^4.1.5",
-    "postcss": "^7.0.27",
+    "postcss": "^7.0.32",
     "prettier": "^2.0.5",
     "remove-markdown": "^0.3.0",
-    "rollup": "^2.7.0",
-    "rollup-plugin-babel": "^4.4.0",
-    "rollup-plugin-terser": "^5.3.0",
-    "sass": "^1.26.3",
-    "sassdoc": "^2.7.1",
-    "sassdoc-theme-herman": "^3.0.2",
+    "rollup": "^2.12.0",
+    "rollup-plugin-terser": "^6.1.0",
+    "sass": "^1.26.7",
+    "sassdoc": "^2.7.2",
+    "sassdoc-theme-herman": "^3.1.0",
     "slugify": "^1.4.0",
     "striptags": "^3.1.1",
-    "stylelint": "^13.3.3",
+    "stylelint": "^13.5.0",
     "stylelint-config-prettier": "^8.0.1",
     "stylelint-config-recommended": "^3.0.0",
     "stylelint-prettier": "^1.1.2",
-    "stylelint-scss": "^3.17.1",
+    "stylelint-scss": "^3.17.2",
     "truncate-html": "^1.0.3",
     "typogr": "^0.6.8"
   },
